{"schema":"decodex/commit/1","summary":"Split contract and operator status models","authority":"manual"}
